Add optional label "-l" for hosts to see which protocol expires
Up to now, checking different protocols on the same hostname in a row gave
cron messages like these:
> CRITICAL: Cert example.org will expire on: 2021-11-27
> CRITICAL: Cert example.org will expire on: 2021-11-27

It was not possible to see which services actually had ssl certificate problems.

This commit adds the optional "-l" label parameter that is output after the
hostname.
If no label is given, the label is fetched from /etc/services.

The output in cron jobs is now:
> CRITICAL: Cert example.org (xmpp) will expire on: 2021-11-27
> CRITICAL: Cert example.org (xmpp-server) will expire on: 2021-11-27